Refactor: organize validators
