+++ ex-4-graphql/server/graphql/mutations/user/createUser.js
@@ -3,7 +3,7 @@ const { GraphQLString } = require('graphql');
 const ErrorHandler = require('../../../utils/ErrorHandler');
 const { User: UserModel } = require('../../../models');
 const { User: UserType } = require('../../types');
-const { user: validateUser } = require('../../../validators/auth');
+const validateUser = require('../../../validators/user');
 
 const createUser = {
 	type: UserType,
