--- ex-4-graphql/server/validators/auth/index.js
@@ -1,7 +1,0 @@
-const status = require('./status');
-const user = require('./user');
-
-module.exports = {
-	status,
-	user
-};
