--- ex-4-graphql/server/validators/auth/status.js
@@ -1,7 +1,0 @@
-const { body } = require('express-validator');
-
-const status = [
-	body('status').trim().not().isEmpty().withMessage('Status cannot be empty.')
-];
-
-module.exports = status;
