--- ex-4-graphql/server/validators/feed/image.js
@@ -1,9 +1,0 @@
-const { body } = require('express-validator');
-
-const image = [
-	body('image')
-		.custom((value, { req }) => req.file !== void 0)
-		.withMessage('No image provided.')
-];
-
-module.exports = image;
