--- ex-4-graphql/server/validators/feed/index.js
@@ -1,9 +1,0 @@
-const image = require('./image');
-const post = require('./post');
-
-const validators = {
-	image,
-	post
-};
-
-module.exports = validators;
